feat: add workflows tabs
