+++ src/assets/css/tailwind.css
@@ -104,7 +104,13 @@ pre {
   &::-webkit-scrollbar-track {
     background: transparent;
   }
+
+  &.scroll-xs::-webkit-scrollbar {
+    width: 5px;
+    height: 5px;
+  }
 }
+
 
 .tippy-box[data-theme~='tooltip-theme'] {
   @apply px-2 py-1 bg-gray-900 dark:bg-gray-200 dark:text-black text-sm text-gray-200 rounded-md;
